requirements.txt
numpy
pandas
scipy
scikit-learn
tensorflow
opencv-python
matplotlib
pyyaml

